Factor out common nodes and grammars
